requirements.txt
openai
Pillow
requests
streamlit